requirements.txt
Flask
google-cloud-storage
firebase
firebase_admin
python-dotenv